+++ src/main/java/roomescape/service/ReservationService.java
@@ -8,8 +8,8 @@ import roomescape.domain.Member;
 import roomescape.domain.ReservationTheme;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Reservation;
-import roomescape.dto.ReservationRecipe;
-import roomescape.dto.ReservationResponse;
+import roomescape.service.dto.ReservationRecipe;
+import roomescape.service.dto.ReservationResponse;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationThemeRepository;
