+++ src/main/java/roomescape/service/dto/AdminReservationRequest.java
@@ -1,9 +1,9 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
-public record ReservationRecipe(
+public record AdminReservationRequest(
         long memberId,
         @DateTimeFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
