+++ src/main/java/roomescape/service/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public record LoginRequest(
         String email,
