+++ src/main/java/roomescape/service/dto/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public record MemberRegisterRequest(
         String email,
