+++ src/main/java/roomescape/service/dto/ReservationThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public record ReservationThemeRequest(String name, String description, String thumbnail) {
 }
