+++ src/main/java/roomescape/service/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 import java.time.LocalTime;
 import org.springframework.format.annotation.DateTimeFormat;
