+++ src/test/java/roomescape/service/MemberServiceTest.java
@@ -9,9 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import roomescape.domain.Member;
-import roomescape.dto.MemberRegisterRequest;
-import roomescape.dto.MemberRegisterResponse;
-import roomescape.dto.MemberResponse;
+import roomescape.service.dto.MemberRegisterRequest;
+import roomescape.service.dto.MemberRegisterResponse;
+import roomescape.service.dto.MemberResponse;
 
 @SpringBootTest
 class MemberServiceTest {
